requirements.txt
fastapi
uvicorn
python-dotenv
requests
sentence-transformers
faiss-cpu
streamlit
numpy
pytz
apscheduler
